usb: misc: starting USBIO driver stack
USBIO driver stack allows the Host to extend the number of IOs available.
It adds IOs exposed by a USB device that complies with the USBIO Bridge
Protocol.

The USBIO stack consists of three drivers:
- USBIO Bridge driver that talks to the USB device exposing its IOs
- USBIO GPIO and I2C drivers that register as gpiochip and i2c adapter
respectively.

Here we add the USBIO Bridge driver in charge of talking to the USB device
firmware. It execute control commands through control transfer pipe.

Note: This USBIO Bridge driver implementation use auxbus framework. There
is concern with this solution due to the behavior seen on the firmware of
some supported usb devices, hence, issues that could emerge should be
addressed by the framework.

Signed-off-by: Israel Cepeda <israel.a.cepeda.lopez@intel.com>
Signed-off-by: Hans de Goede <hansg@kernel.org>