usb: misc: Starting USBIO driver stack
USBIO driver stack allows the Host to extend the number of IOs available.
It adds IOs exposed by a USB device that complies with the USBIO Bridge
Protocol.

The USBIO stack consists of three drivers:
- USBIO Bridge driver that talks to the USB device exposing its IOs
- USBIO GPIO and I2C drivers that register as gpiochip and i2c adapter
  respectively.

Here we add the USBIO Bridge driver in charge to talk to the USB device
firmware. It execute control commands through control transfer pipe.

Signed-off-by: Israel Cepeda <israel.a.cepeda.lopez@intel.com>